made resuable ordanalize method
