+++ src/utils/helpers.js
@@ -62,6 +62,19 @@ export function formatTime(date) {
   })
 }
 
+export function ordinalize(num) {
+  const n = Number(num)
+  if (!Number.isFinite(n) || n === 0) return '—'
+  const mod100 = n % 100
+  if (mod100 >= 11 && mod100 <= 13) return `${n}<sup>th</sup>`
+  switch (n % 10) {
+    case 1: return `${n}<sup>st</sup>`
+    case 2: return `${n}<sup>nd</sup>`
+    case 3: return `${n}<sup>rd</sup>`
+    default: return `${n}<sup>th</sup>`
+  }
+}
+
 export function formatDuration(duration) {
   const totalSeconds = Math.floor(duration / 1000)
